[PSCALE-220239] Remove get_available_subnet.sh from TF-azurerm-PS
Note: TF-azurerm-PS   == terraform-azurerm-powerscale

The get_available_subnet.sh script was used during the PoC work to quickly get available IPs for 
the cluster subnets. Since the terraform-azurerm-powerscale repo is a shippable repo and the 
script will not be shipped with the repo, we need to remove it from the repo.

This patch does the following:
- Remove the  get_available_subnet.sh  from the terraform-azurerm-powerscale repo
- Update terraform-azurerm-powerscale to use subnet CIDRs
- Some minor refactoring